@@ -11,9 +11,10 @@
     <module>chill-script</module>
     <module>chill-web</module>
     <module>autom8</module>
+      <module>chill-job</module>
   </modules>
   <properties>
-    <maven.compiler.source>11</maven.compiler.source>
-    <maven.compiler.target>11</maven.compiler.target>
+    <maven.compiler.source>17</maven.compiler.source>
+    <maven.compiler.target>17</maven.compiler.target>
   </properties>
 </project>
