@@ -3,5 +3,5 @@ package chill.script.parser;
 import chill.script.expressions.Expression;
 
 public interface ExpressionParser {
-    Expression parse(ChillScriptParser parser);
+    Expression parseExpression(ChillScriptParser parser);
 }
